web: Split off installRuffle, PublicAPI and SourceAPI to Setup
