+++ web/packages/core/src/index.ts
@@ -4,11 +4,10 @@
  * Types should only be exported here if they are intended to be part of that public API, not internal.
  */
 
-export * from "./public-api";
+export * as Setup from "./public/setup";
 export * as Config from "./public/config";
 export * from "./build-info";
 export * from "./movie-metadata";
-export * from "./install";
 export * from "./public/flash";
 export * from "./public/legacy";
 export * from "./public/player";
