+++ web/packages/core/src/public/setup/index.ts
@@ -1,0 +1,3 @@
+export * from "./public-api";
+export * from "./source-api";
+export * from "./install";
